+++ examples/multi_file/utils.rs
@@ -1,0 +1,4 @@
+/// Displays a "hello!" message
+pub fn print_hello() {
+    println!("hello!")
+}