+++ examples/run_example.rs
@@ -12,8 +12,23 @@ fn main() {
     process_input(Args {
         name: Some("hello_world".into()),
         args: Vec::from(args.map(String::from)),
-        prompt_args: false,
-        verbose: true,
+        input_args: false,
+        cargo: cargo_options::CommonOptions {
+            // enable verbose output
+            verbose: 1,
+            // uncomment to build in `--release` mode
+            // release: true,
+            jobs: Some(2),
+            profile: Some("release".to_owned()),
+            features: vec!["__feature-1".to_owned()],
+            no_default_features: true,
+            ignore_rust_version: true,
+            color: Some("always".to_owned()),
+            locked: true,
+            offline: true,
+            config: vec!["my_cfg='test'".to_owned()],
+            ..Default::default()
+        },
         ..Default::default()
     })
     .unwrap();