+++ src/constants.rs
@@ -21,7 +21,14 @@ pub const SETTINGS_TOML: &str = "settings.toml";
 
 /// The `cargo` command
 pub const CARGO_CMD: &str = "cargo";
-/// The `echo` command, present in both Linux/Windows environments
-pub const ECHO_CMD: &str = "echo";
 /// The `fzf` (fuzzy finder) command
 pub const FZF_CMD: &str = "fzf";
+
+/// Newline character as bytes
+pub const NL: &u8 = &b'\n';
+
+/// Space character as bytes
+pub const SPACE: &u8 = &b' ';
+
+#[cfg(any(target_os = "windows", target_arch = "wasm32"))]
+pub(crate) const INVALID_UTF8: &str = "unexpected invalid UTF-8 code point";